+++ src/features/notification/model/index.ts
@@ -1,0 +1,2 @@
+export * from './notification.query'
+export { default as useNotification } from './use-notification'
