+++ src/features/notification/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as NotificationList } from './notification-list'
+export { default as NotificationBell } from './notification-bell'
